Claude/playerbot cleanup analysis 01 cc h sn awqm7 w9z ve uu jmf4h

+++ src/modules/Playerbot/Core/DI/Interfaces/IFarmingCoordinator.h
@@ -20,7 +20,7 @@ namespace Playerbot
 
 // Forward declarations
 enum class FarmingSessionType : uint8;
-enum class ProfessionType : uint8;
+enum class ProfessionType : uint16;
 struct FarmingZoneInfo;
 struct FarmingSession;
 struct FarmingCoordinatorProfile;
